Added GroupOwnerID column to the Groups table in CustomDatabaseOpenHelper.java.
Added ScrapbookGroups table in CustomDatabaseOpenHelper.java, used to associate scrapbooks with Groups

Added createGroup, joinGroup, and postToGroup methods to the GroupDAO interface.

Added getGroupDAO methods to the BaseFragment, MainActivity and Scrumble classes.

Added (though currently empty) a GroupDAOImplementation class, which implements the GroupDAO interface.

Added a second constructor to the Group class, which takes the name of the group as an argument.

Created a "fragment_new_group" fragment, with "NewGroupFragment" class associated with it.